Consolidate GDS tests into test_cuda_gds.py
